google-adk/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
litellm
google-adk
